✨feat: Improve init.sh

+++ init.sh
@@ -5,29 +5,41 @@ source "$PWD/.env"
 
 echo "INITIALIZED ENV"
 
-ln -s "$PWD/Invitation" "$LOR_MOD_DIR/$ID"
+if [ ! -d "$LOR_MOD_DIR/$ID" ]; then
+	ln -s "$PWD/Invitation" "$LOR_MOD_DIR/$ID"
 
-echo "SYMLINK CREATED"
+	echo "SYMLINK CREATED"
+else
+	echo "SYMLINK ALREADY CREATED"
+fi
 
-mkdir -p "$PWD/src/libs/"
-cp "$LOR_DATA_DIR/System.dll" "$PWD/src/libs/System.dll"
-cp "$LOR_DATA_DIR/System.Core.dll" "$PWD/src/libs/System.Core.dll"
-cp "$LOR_DATA_DIR/mscorlib.dll" "$PWD/src/libs/mscorlib.dll"
-cp "$LOR_DATA_DIR/Assembly-CSharp.dll" "$PWD/src/libs/Assembly-CSharp.dll"
+if [ ! -d "$PWD/src/libs/" ]; then
+	mkdir -p "$PWD/src/libs/"
+	cp "$LOR_DATA_DIR/System.dll" "$PWD/src/libs/System.dll"
+	cp "$LOR_DATA_DIR/System.Core.dll" "$PWD/src/libs/System.Core.dll"
+	cp "$LOR_DATA_DIR/mscorlib.dll" "$PWD/src/libs/mscorlib.dll"
+	cp "$LOR_DATA_DIR/Assembly-CSharp.dll" "$PWD/src/libs/Assembly-CSharp.dll"
 
-echo "LOCAL ASSEMBLIES SUCCESSFULLY COPIED"
+	echo "LOCAL ASSEMBLIES SUCCESSFULLY COPIED"
+else
+	echo "LOCAL ASSEMBLIES ALREADY COPIED"
+fi
 
-mv "$PWD/src/.csproj" "$PWD/src/$ID.csproj"
+if [ -e "$PWD/src/.csproj" ]; then
+	mv "$PWD/src/.csproj" "$PWD/src/$ID.csproj"
 
-sed -i "s/__MODNAME__/$ID/g" "$PWD/src/Init.cs"
-sed -i "s/__MODNAME__/$ID/g" "$PWD/src/PatchClass.cs"
-sed -i "s/__MODNAME__/$ID/g" "$PWD/src/$ID.csproj"
-sed -i "s/__MODNAME__/$ID/g" "$PWD/Invitation/StageModInfo.xml"
-sed -i "s/__TITLE__/$WORKSHOP_TITLE/g" "$PWD/Invitation/StageModInfo.xml"
-sed -i "s/__DESC__/$WORKSHOP_DESC/g" "$PWD/Invitation/StageModInfo.xml"
-sed -i "s/__THUMBNAIL__/$WORKSHOP_THUMBNAIL/g" "$PWD/Invitation/StageModInfo.xml"
+	sed -i "s/__MODNAME__/$ID/g" "$PWD/src/Init.cs"
+	sed -i "s/__MODNAME__/$ID/g" "$PWD/src/PatchClass.cs"
+	sed -i "s/__MODNAME__/$ID/g" "$PWD/src/$ID.csproj"
+	sed -i "s/__MODNAME__/$ID/g" "$PWD/Invitation/StageModInfo.xml"
+	sed -i "s/__TITLE__/$WORKSHOP_TITLE/g" "$PWD/Invitation/StageModInfo.xml"
+	sed -i "s/__DESC__/$WORKSHOP_DESC/g" "$PWD/Invitation/StageModInfo.xml"
+	sed -i "s/__THUMBNAIL__/$WORKSHOP_THUMBNAIL/g" "$PWD/Invitation/StageModInfo.xml"
 
-echo "REPLACED ALL TEMPLATE SYNTAX"
+	echo "REPLACED ALL TEMPLATE SYNTAX"
+else
+	echo "ALL TEMPLATE ALREADY REPLACED"
+fi
 
 chmod +x "$PWD/dev.sh"
 chmod +x "$PWD/build.sh"
